fix: 发布 v0.3.0.4，修复交替斜向奔跑并放开时序设置

+++ native/engine_api.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <windows.h>
 extern "C" {
-void* AF_Start(const unsigned*, unsigned, unsigned, unsigned, unsigned, HWND, ULONG_PTR);
+void* AF_Start(const unsigned*, unsigned, unsigned long long, unsigned long long, unsigned, HWND, ULONG_PTR);
 int AF_Stop(void*);
 int AF_Physical(void*, unsigned, unsigned);
 unsigned AF_Error(void*);
